requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
tqdm>=4.65
torch>=2.1
torchvision>=0.16
dreamsim>=0.1.2
transformers>=4.35
open_clip_torch>=2.23
monai>=1.3
nibabel>=5.1
Pillow>=10.0
scikit-image>=0.21
psutil>=5.9
